% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{breast}
\alias{breast}
\title{breast}
\format{
A list containing a data frame with 7616 rows and 8 variables:
}
\usage{
breast
}
\description{
cleaned and ready data frame for breast cancer
}
\keyword{datasets}
