% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cancerData}
\alias{cancerData}
\title{cancerData}
\format{
A list containing three dataframes or lists
}
\usage{
cancerData
}
\description{
three data frames, prostate, breast and trachea
}
\keyword{datasets}
