% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cattle.pheno.R
\docType{data}
\name{cattle.pheno}
\alias{cattle.pheno}
\title{Phenotypic information of cattle dataset}
\description{
A matrix with a dimension of 2,500 x 6, which includes one phenotype, 
two fixed covariates, and pedigree information.
}
\author{
Haipeng Yu and Gota Morota 

Maintainer: Haipeng Yu \email{haipengyu@vt.edu}
}
\keyword{datasets}
