% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cattle.W.R
\docType{data}
\name{cattle.W}
\alias{cattle.W}
\title{Marker information of cattle dataset}
\description{
A 2,500 by 10,000 matrix contains marker information.
}
\author{
Haipeng Yu and Gota Morota 

Maintainer: Haipeng Yu \email{haipengyu@vt.edu}
}
\keyword{datasets}
