% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CattleSIM.R
\docType{data}
\name{CattleSIM}
\alias{CattleSIM}
\title{Cattle dataset with simulated management units}
\usage{
data(CattleSIM)
}
\description{
The cattle dataset is simulated with QMSim software (Sargolzaei and Schenkel, 2009).   
This dataset includes 2,500 individuals across six generations (from founder to generation 5), 
each with 10,000 single nucleotide polymorphisms spread over 29 autosomes. Single phenotype with heritability of 0.6 was simulated.
In addition, six fixed effects including one cluster effect and 5 management unit simulation scenarios were simulated. 
Fixed effect of the cluster is simulated using the K-medoid algorithm (Kaufman and Rousseeuw, 1990) to assign 2,500 individuals into eight clusters. 
MUSC1 is simulated by randomly allocating eight clusters into two sets, which is regarded as the least connected design. From MUSC2 to 5,  
randomly sampled individuals (i.e., 140, 210, 280, and 350) were exchanged between the two sets in MUSC1 to steadily increase 
the degree of relatedness.
}
\examples{
# Load cattle dataset
data(CattleSIM)

# Check management units simulation
str(CattleSIM)

}
\references{
Sargolzaei, M., and F. S. Schenkel. 2009. Qmsim: a large-scale
  genome simulator for livestock. Bioinformatics 25:680–681. doi:10.1093/bioinformatics/btp045

Kaufman, L. and P. Rousseeuw. 1990. Finding groups in data: 
an introduction to cluster analysis. John Wiley and Sons, New York.
}
\author{
Haipeng Yu and Gota Morota 

Maintainer: Haipeng Yu \email{haipengyu@vt.edu}
}
\keyword{datasets}
