% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{checkValidity}
\alias{checkValidity}
\title{check data available/complete}
\usage{
checkValidity(df, registry)
}
\arguments{
\item{df}{the dataframe to be checked}

\item{registry}{the registry dataframe}
}
\description{
using the registry, check which rows are available/complete
}
