% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{cleanAndCombine}
\alias{cleanAndCombine}
\title{clean and combine files}
\usage{
cleanAndCombine(incFile, popFile, mortFile)
}
\arguments{
\item{incFile}{name of incidence file}

\item{popFile}{name of population file}

\item{mortFile}{name of mortality file}
}
\value{
a list containing all three data frames
}
\description{
cleans and combines all three files, combining them to data frames and checks validity
}
