% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{collapseAgeGroups}
\alias{collapseAgeGroups}
\title{Collapse Age groups}
\usage{
collapseAgeGroups(df)
}
\arguments{
\item{df}{dataframe}
}
\description{
collapse age groups in population and incidence data,
as mortality data often has 0-14 yrs as one single group
}
