% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\docType{data}
\name{completeStates}
\alias{completeStates}
\title{complete states}
\format{
An object of class \code{character} of length 8.
}
\usage{
completeStates
}
\description{
when aggregating data over states, only these states are used.
These have complete data from 2001 to 2014.
}
\keyword{datasets}
