% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeA.R
\name{computeA}
\alias{computeA}
\title{Computation of numerator relationship matrix.}
\usage{
computeA(Progeny, Sire, Dam)
}
\arguments{
\item{Progeny}{a vector of sorted animals}

\item{Sire}{a vector of Sires according to progenies, where the missing Sire is coded as 0}

\item{Dam}{a vector of Dams according to progenies, where the missing Dam is coded as 0}
}
\value{
a n by n numerator relationship matrix.
}
\description{
Use sorted pedigree to calculate numerator relationship matrix. The missing parents are coded as 0.
}
\examples{
computeA()

}
