% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortModule.R
\name{convertInterval}
\alias{convertInterval}
\title{convert agegroups}
\usage{
convertInterval(string)
}
\arguments{
\item{string}{age group string to be converted to an int}
}
\description{
convert agegroups to median integer
}
