% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataChoiceModule.R
\name{dataChoiceServer}
\alias{dataChoiceServer}
\title{The server module for choosing dataframes.}
\usage{
dataChoiceServer(id)
}
\arguments{
\item{id}{passed to the module in 'app.R', must match with id passed to UI.}
}
\value{
a dataframe for further use
}
\description{
The server module for choosing dataframes.
}
