% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataChoiceModule.R
\name{dataChoiceUI}
\alias{dataChoiceUI}
\title{The UI module for choosing dataframes.}
\usage{
dataChoiceUI(id)
}
\arguments{
\item{id}{passed to the module in 'app.R', must match with id passed to server.}
}
\description{
The UI module for choosing dataframes.
}
