% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{fed_total}
\alias{fed_total}
\title{fix data for single federal state}
\usage{
fed_total(incFile, popFile, mortFile, sheet)
}
\arguments{
\item{incFile}{incidence file name}

\item{popFile}{population file name}

\item{mortFile}{mortality file name}

\item{sheet}{abbrevation for federal state, specifying excel sheet}
}
\description{
gets incidence, population and mortality data for singel federal state
and mereges together using mergeData
}
