% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCA.R
\docType{package}
\name{GCA-package}
\alias{GCA}
\alias{GCA-package}
\title{\strong{GCA}: Genetic connectedness analysis}
\description{
An R package for genetic connectedness analysis across units using pedigree and genomic data.
}
\section{Estimates of connectedness}{

gcm()
}

\section{Estimate variance components}{

varcomp()
}

\section{Compute genomic relationship matrix}{

computeG()
}

\section{Compute numerator relationship matrix}{

computeA()
}

\seealso{
Useful links:
\itemize{
  \item <\url{https://haipengu.github.io/}>
}
}
\author{
Haipeng Yu \email{haipengyu@vt.edu}

Gota Morota \email{morota@vt.edu}
}
\keyword{internal}
