% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogramModule.R
\name{histogramPlotter}
\alias{histogramPlotter}
\title{plots incidence and mortality}
\usage{
histogramPlotter(df)
}
\arguments{
\item{df}{dataframe}
}
\description{
plots histogram for each federal state. x-variable is agegroup and y-variable is
incidence- and mortality rate
}
