% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{meltAndClean}
\alias{meltAndClean}
\title{Melt data}
\usage{
meltAndClean(df, type)
}
\arguments{
\item{df}{two dimensional array}

\item{type}{incidence or mortality data}
}
\description{
melt two dimensional array into dataframe, each entry gets a unique row
}
