% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{mergeData}
\alias{mergeData}
\title{merge data}
\usage{
mergeData(inc, pop, mort, sheet)
}
\arguments{
\item{inc}{list with incidence data frame}

\item{pop}{list with population data frame}

\item{mort}{list with mortality data frame}

\item{sheet}{abbrevation specifying federal state}
}
\description{
merge data for incidence, population and mortality for one specific federal state
}
