% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageGroupAnalysis.R
\name{plotAgGrid}
\alias{plotAgGrid}
\title{Plot age-stratified plot}
\usage{
plotAgGrid(df, fstate)
}
\arguments{
\item{df}{the cleaned dataframe, for example from the provided list cancerData}

\item{fstate}{specify federalstate, two-letters abbrevations, see variable states}
}
\value{

}
\description{
Plots age-stratified plot, both incidence and mortality rates.
For one specific state or all states aggregated.
}
\examples{
prostateDf <- cancerData[['prostate']][[1]]
prostateDf <- removeNans(prostateDf)
df <- prepareData(df=prostateDf,fstate='SL')
plotAgGrid(df=df,fstate='SL')
}
