% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageGroupAnalysis.R
\name{plotAgMortPerInc}
\alias{plotAgMortPerInc}
\title{plot age group plot}
\usage{
plotAgMortPerInc(df, fstate)
}
\arguments{
\item{df}{dataframe}

\item{fstate}{federal state}
}
\description{
figure with separate plot for each data frame, plotting mortality divided by incidence against period
}
