% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageStandardizedAnalysis.R
\name{plotAsis}
\alias{plotAsis}
\title{plot age standardized analysis}
\usage{
plotAsis(df, fstate)
}
\arguments{
\item{df}{dataframe, prepared using tweakAsisData}

\item{fstate}{specifying federal state}
}
\description{
plots age standardized analysis
}
\examples{
prostate <- cancerData[['prostate']][[1]]
data <- tweakAsisData(prostate,'SL')
plotAsis(data,'SL')
}
