% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoAnalysisModule.R
\name{plotGeo}
\alias{plotGeo}
\title{display incidence plots arranged as Germany}
\usage{
plotGeo(df)
}
\arguments{
\item{df}{dataframe}
}
\description{
display incidence for each state, arranged as germany.
}
