% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoAnalysisModule.R
\name{plotGeoMort}
\alias{plotGeoMort}
\title{display mortality plots arranged as Germany}
\usage{
plotGeoMort(df)
}
\arguments{
\item{df}{dataframe}
}
\description{
display incidence for each state, arranged as germany.
}
