% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoAnalysisModule.R
\name{plotter}
\alias{plotter}
\title{display plots arranged as Germany}
\usage{
plotter(df, varxString, varyString)
}
\arguments{
\item{df}{dataframe}

\item{varxString}{x-variable, used for xlabel}

\item{varyString}{y-variable, used for ylabel}
}
\description{
display incidence or mortality for each state, arranged as germany.
}
