% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{prepareData}
\alias{prepareData}
\title{prepareData}
\usage{
prepareData(df, fstate = "ALL*")
}
\arguments{
\item{df}{dataframe to prepare}

\item{fstate}{federalstate, defaults to ALL*}
}
\value{
cleaned dataframe
}
\description{
prepares data by choosing fedState or summing over fedState. Also calculates values per capita
}
