% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{read_sheet}
\alias{read_sheet}
\title{read sheet}
\usage{
read_sheet(name, type, sheet)
}
\arguments{
\item{name}{name of file}

\item{type}{specifying whether its incidence, mortality or population data}

\item{sheet}{sheet abbrevation specifying federal state, for example 'BB'}
}
\description{
read sheet for one Federal state
}
