% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{registry}
\alias{registry}
\title{registry}
\format{
A data frame
}
\usage{
registry
}
\description{
data frame containing information of the state of data, available and/or complete
}
\keyword{datasets}
