% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{removeNans}
\alias{removeNans}
\title{Removing NaN values for complete column}
\usage{
removeNans(df)
}
\arguments{
\item{df}{The data frame whose NaN values are set to False}
}
\value{
The dataframe without NaN.
}
\description{
NaNs in the complete/available columns are set to FALSE.
}
\examples{
prostate <- cancerData[['prostate']][[1]]
removeNans(prostate)
}
