% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{splitByGender}
\alias{splitByGender}
\title{split by gender}
\usage{
splitByGender(tibble)
}
\arguments{
\item{tibble}{Takes in tibble subject to splitting}
}
\value{
returns list with two dataframes, one for each gender
}
\description{
splits data frames by gender
}
