% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trachea}
\alias{trachea}
\title{trachea}
\format{
A list containing two data frames, one for for women and for men
}
\usage{
trachea
}
\description{
cleaned and ready data frame for lung cancer
}
\keyword{datasets}
