% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageStandardizedAnalysis.R
\name{tweakAsisData}
\alias{tweakAsisData}
\title{Prepare data for age standardized analysis}
\usage{
tweakAsisData(data, fstate)
}
\arguments{
\item{data}{data frame for preparation}

\item{fstate}{specifying federal state}
}
\value{
prepared data frame, ready for age standardized analysis
}
\description{
Prepares data for age standardized analysis.
Pretty similar to prepareData(), except for a few things.
Makes two new columns for the reweigheted incidence and mortality rates
}
