% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortModule.R
\name{tweakCohortData}
\alias{tweakCohortData}
\title{prepare data for cohort analysis}
\usage{
tweakCohortData(data, fstate)
}
\arguments{
\item{data}{dataframe}

\item{fstate}{abbrevation for federal state}
}
\description{
data preparation similar to PrepareData(), but with a few extra steps
}
