% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogramModule.R
\name{tweakHistogramData}
\alias{tweakHistogramData}
\title{prepare data for histogram plot}
\usage{
tweakHistogramData(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
dataframe ready for plotting
}
\description{
prepares data frame for histogram plotting.
Selects only complete data and then sums over agegroups
}
