% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\docType{data}
\name{weights}
\alias{weights}
\title{european standard weights}
\format{
An object of class \code{list} of length 16.
}
\usage{
weights
}
\description{
european standard weight for age standardized analysis
}
\keyword{datasets}
