% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeA.R
\name{computeA}
\alias{computeA}
\title{Compute numerator relationship matrix.}
\usage{
computeA(Progeny, Sire, Dam)
}
\arguments{
\item{Progeny}{A numeric vector of sorted progenies.}

\item{Sire}{A numeric vector of sires according to progenies in \code{Progeny} column.}

\item{Dam}{A numeric vector of dams according to progenies in \code{Progeny} column.}
}
\value{
A n by n numerator relationship matrix, where n refers to total number of individuals in \code{Progeny} column.
}
\description{
Use sorted pedigree to calculate a numerator relationship matrix based on a tabular method, where the individuals in \code{Progeny} must be ordered genealogically. 
 The missing \code{Sire} and \code{Dam} are coded as 0.
}
\examples{
# Load GCA package
library(GCA)

# Import cattle dataset
data(GCcattle)

# Pedigree information
head(cattle.pheno)

# Construct numerator relationship matrix
NRM <- computeA(Progeny = cattle.pheno$Progeny, Sire = cattle.pheno$Sire, 
		Dam = cattle.pheno$Dam)
}
\author{
Haipeng Yu and Gota Morota 

Maintainer: Haipeng Yu \email{haipengyu@vt.edu}
}
