% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeG.R
\name{computeG}
\alias{computeG}
\title{Compute genomic relationship matrix}
\usage{
computeG(snpmatrix, maf = 0.05, impute = "rbinom", method = "G1")
}
\arguments{
\item{snpmatrix}{A marker matrix with the dimension of n by m, where the elements are coded as 0, 1, 2, or NA, 
where n and m are the total number of individuals and markers, respectively.}

\item{maf}{A minor allele frequency cutoff for quality control. The default minor allele frequency is 0.05.}

\item{impute}{Perform genotype imputation for missing markers if applicable. Two methods of 'mean' and 'rbinom' are available, 
where the 'mean' imputes missing markers using mean and 'rbinom' imputes the missing markers by random sampling from 
a binomial distribution. The default method is 'rbinom'. This argument will be ignored if the \code{snpmatrix} does not include any missing markers.}

\item{method}{A type of genomic relationship matrix including 'G1' and 'G2' (VanRaden 2008). The default method is 'G1'.}
}
\value{
An n by n additive genomic relationship matrix.
}
\description{
Use single nucleotide polymorphisms markers to derive an additive genomic relationship matrix. Missing markers are allowed, but should be coded as NA.
}
\examples{
# Load cattle data
data(GCcattle)

# Marker information
str(cattle.W)

# Compute genomic relationship matrix
G <- computeG(cattle.W, maf = 0.05, impute = 'rbinom', method = 'G1')
}
\references{
\emph{VanRaden, P.M., 2008. Efficient methods to compute genomic predictions. Journal of dairy science, 91(11), pp.4414-4423.}
}
\author{
Haipeng Yu and Gota Morota 

Maintainer: Haipeng Yu \email{haipengyu@vt.edu}
}
