% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCA.R
\docType{package}
\name{GCA-package}
\alias{GCA}
\alias{GCA-package}
\title{\strong{GCA}: Genetic connectedness analysis}
\description{
An R package for genetic connectedness analysis across units using pedigree and genomic data.
}
\details{
The GCA package encompasses numerous connectedness statistics, which could be labeled as two groups, 
by reference to connectedness based on prediction error variance (PEV) and variance of unit effect estimates (VE).
The PEV-derived metrics include prediction error variance of differences (PEVD), coefficient of determination (CD), 
and prediction error correlation (r). These PEV-derived metrics can be summarized at the unit level as the average PEV within
and across units (GrpAve), average PEV of all pairwise differences between individuals across units (IdAve), 
or using a contrast vector (Contrast). VE-derived metrics comprise variance of differences in management unit effects (VED), 
coefficient of determination of VED (CDVED), and connectedness rating (CR). Three correction factors accounting for the number 
of fixed effects can be applied for each VE-derived metric. These include non-correction (0), correction of unit effect (1), 
and correction of two or more fixed effects (2). The core function of GCA is integrated with C++ to improve computational 
efficiency using the Rcpp package (Eddelbuettel and François 2011). The details of these connectedness statistics can be found
in Yu and Morota 2019.
}
\section{Available functions in GCA package}{

\itemize{
  \item computeA(): Computation of numerator relationship matrix. 
  \item computeG(): Computation of genomic relationship matrix.
  \item gca(): Measures of genetic connectedness.
  \item varcomp(): Estimates of variance components using eigenvalues and eigenvectors. 
}
}

\references{
\emph{Eddelbuettel D, François R (2011). Rcpp: Seamless R and C++ Integration. Journal of Statistical Software, 40(8), 1–18.}

\emph{Yu H and Morota G. 2019. GCA: An R Package for Genetic Connectedness Analysis Using Pedigree and Genomic Data.}
}
\author{
Haipeng Yu and Gota Morota 

Maintainer: Haipeng Yu \email{haipengyu@vt.edu}
}
\keyword{internal}
