% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCcattle.R
\docType{data}
\name{GCcattle}
\alias{GCcattle}
\title{Cattle dataset}
\format{
\itemize{
  \item cattle.pheno: A matrix with a dimension of 2500 x 6, which includes one phenotype, two fixed covariates and pedigree information.
  \item catlle.W: A 2500 by 10,000 matrix, which contains marker information.
}
}
\usage{
data(GCcattle)
}
\description{
The cattle dataset is simulated with QMSim software (Sargolzaei and Schenkel, 2009).   
This dataset includes 2,500 individuals across six generations (from founder to generation 5), 
each with 10,000 single nucleotide polymorphisms spread over 29 autosomes. Single phenotype with heritability of 0.6 was simulated.
Two fixed covariates of sex and unit are available.
}
\examples{
# Load cattle dataset
data(GCcattle)

# Check phenotype, fixed covariates and pedigree information
str(cattle.pheno)

# Check marker information
str(cattle.W)
}
\references{
Sargolzaei, M., and F. S. Schenkel. 2009. Qmsim: a large-scale
  genome simulator for livestock. Bioinformatics 25:680–681. doi:10.1093/bioinformatics/btp045
}
\author{
Haipeng Yu and Gota Morota 

Maintainer: Haipeng Yu \email{haipengyu@vt.edu}
}
\keyword{datasets}
