% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{getCompleteYears}
\alias{getCompleteYears}
\title{complete years for federal state}
\usage{
getCompleteYears(fstate)
}
\arguments{
\item{fstate}{federal state}
}
\value{
two integers, indicating start and end of complete period
}
\description{
get years for which the specified federal state have complete data
}
\examples{
getCompleteYears('SL')
}
