% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageGroupAnalysis.R
\name{plotAgGrid}
\alias{plotAgGrid}
\title{plot age group plot}
\usage{
plotAgGrid(df, fstate)
}
\arguments{
\item{df}{dataframe}

\item{fstate}{federal state}
}
\description{
figure with separate plot for each data frame, plotting incidence or mortality against period
}
