% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{removeNans}
\alias{removeNans}
\title{Removing NaN values}
\usage{
removeNans(df)
}
\arguments{
\item{df}{The data frame whose Nan values are set to False}
}
\value{
The dataframe
}
\description{
Nans in the complete/available columns are set to FALSE.
}
