% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoAnalysisModule.R
\name{sumOverGroups}
\alias{sumOverGroups}
\title{summing over groups, specific for geoAnalysis plot}
\usage{
sumOverGroups(df, varx, vary)
}
\arguments{
\item{df}{dataframe}

\item{varx}{x-variable}

\item{vary}{y-variable}
}
\description{
summing over groups, specific for geoAnalysis plot
}
