% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageStandardizedAnalysis.R
\name{tweakAsisData}
\alias{tweakAsisData}
\title{Data preparation II}
\usage{
tweakAsisData(data, fstate)
}
\arguments{
\item{data}{data frame for preparation}

\item{fstate}{specifying federal state}
}
\value{
prepared data frame, ready for further analyses
}
\description{
Prepares data for age standardized analysis.
Resembles prepareData(). Final dataframe only has period, incRate and mortRate
}
\examples{
prostate <- cancerData[['prostate']][[1]]
data <- tweakAsisData(prostate,'SL')
plotAsis(data,'SL')
}
