% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogramModule.R
\name{tweakHistogramData}
\alias{tweakHistogramData}
\title{prepare data for histogram plot}
\usage{
tweakHistogramData(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
dataframe ready for plotting
}
\description{
prepares data frame for histogram plotting.
Selects only complete data and then sums over agegroups. Returns df
with federal state, incidence, mortality, population, and two
columns for coloring in histogram
}
\examples{
prostate <- removeNans(cancerData[['prostate']][[1]])
data <- tweakHistogramData(prostate)
histogramPlotter(data)
}
