% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varcomp.R
\name{varcomp}
\alias{varcomp}
\title{Estimates of variance componets.}
\usage{
varcomp(y, Evector, Evalue)
}
\arguments{
\item{y}{a n by 1 vector of phenotypes.}

\item{Evector}{a n by n matrix with columns according to eigenvectors.}

\item{Evalue}{a vector contains the n eigenvalues.}
}
\value{
A list of variance components.
}
\description{
Maximum Likelihood estimation of variance components with eigenvectors.
}
\examples{
varcomp()

}
